refactor channel registers
