+++ qiskit/pulse/channels/channel_register.py
@@ -14,8 +14,7 @@ from typing import Union, List
 from qiskit.circuit import Register
 from qiskit.exceptions import QiskitError, QiskitIndexError
 from qiskit.pulse.exceptions import ChannelsError
-from .output_channels import OutputChannel
-from .pulse_channel import PulseChannel
+from .pulse_channel import PulseChannel, AcquireChannel, SnapshotChannel
 
 logger = logging.getLogger(__name__)
 
@@ -56,9 +55,9 @@ class ChannelRegister(Register):
         """
         if not isinstance(key, (int, slice, list)):
             raise QiskitError("expected integer or slice index into register")
-        self.check_range(key)   # TODO: need more checks!
+        self.check_range(key)  # TODO: need more checks!
         if isinstance(key, list):  # list of channel indices
-            if max(key) < len(self):    # TODO: need to move check_range()
+            if max(key) < len(self):  # TODO: need to move check_range()
                 return [self._channels[i] for i in key]
             else:
                 raise QiskitIndexError('register index out of range')
@@ -93,41 +92,22 @@ class ChannelRegister(Register):
         """Make object hashable."""
         return hash((type(self), self.channel_cls, self.name, self.size))
 
-
-class OutputChannelRegister(ChannelRegister):
-    """Implement an output channel register."""
-
-    def __init__(self, channel_cls, size: int, name: str = None, lo_freqs: List[float] = None):
-        """Create a new output channel register.
-        """
-        if not issubclass(channel_cls, OutputChannel):
-            raise ChannelsError("Unknown output channel class: %s", channel_cls.__name__)
-
-        super().__init__(channel_cls, size, name)
-
-        if lo_freqs:
-            self.lo_freqs = lo_freqs
-        else:
-            self.lo_freqs = [None] * size
-
-    def __eq__(self, other):
-        """Two channel registers are the same if they are of the same type
-         have the same channel class, name and size.
-
-        Args:
-            other (OutputChannelRegister): other OutputChannelRegister
-
-        Returns:
-            bool: are self and other equal.
-        """
-        if type(self) is type(other) and \
-                self.channel_cls == other.channel_cls and \
-                self.name == other.name and \
-                self.size == other.size and \
-                self.lo_freqs == other.lo_freqs:
-            return True
-        return False
-
-    def __hash__(self):
-        """Make object hashable."""
         return hash((type(self), self.channel_cls, self.name, self.size, self.lo_freqs))
+
+
+class AcquireChannelRegister(ChannelRegister):
+    """Acquire channel register."""
+
+    def __init__(self, size: int, name: str = None):
+        """Create a new aquire channel register.
+        """
+        super().__init__(AcquireChannel, size, name)
+
+
+class SnapshotChannelRegister(ChannelRegister):
+    """Snapshot channel register."""
+
+    def __init__(self, size: int, name: str = None):
+        """Create a new snapshot channel register.
+        """
+        super().__init__(SnapshotChannel, size, name)
