+++ qiskit/pulse/channels/output_channel.py
